Add files via upload
This repository contains a custom implementation of "ISAC from the Sky: UAV Trajectory Design for Joint Communication and Target Localization". The implementation in this repository is based on this paper. This repository is independent of the original paper.

This repository contains a custom implementation of "ISAC from the Sky: UAV Trajectory Design for Joint Communication and Target Localization". The implementation in this repository is based on this paper. This repository is independent of the original paper.

thanks to https://github.com/sAz-G/isac_fts/tree/main